requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-serializer-extensions==2.0.1
gunicorn==20.0.4
hashids==1.3.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
psycopg2==2.8.6
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-decouple==3.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.11
whitenoise==5.2.0
wrapt==1.12.1
